refactor: restructure app and translate docs
