+++ app/__init__.py
@@ -1,0 +1,1 @@
+"""ShiftTradeAV application package."""
