+++ app/pages/__init__.py
@@ -1,0 +1,1 @@
+"""Streamlit pages for ShiftTradeAV."""
